# [hail] Introduce `ArrayFold2` to permit more deforesting