--- a/hail/src/main/scala/is/hail/expr/ir/Copy.scala
+++ b/hail/src/main/scala/is/hail/expr/ir/Copy.scala
@@ -130,6 +130,13 @@ object Copy {
       case ArrayFold(_, _, accumName, valueName, _) =>
         val IndexedSeq(a: IR, zero: IR, body: IR) = newChildren
         ArrayFold(a, zero, accumName, valueName, body)
+      case ArrayFold2(_, accum, valueName, seq, _) =>
+        val ncIR = newChildren.map(_.asInstanceOf[IR])
+        assert(newChildren.length == 2 + accum.length + seq.length)
+        ArrayFold2(ncIR(0),
+          accum.indices.map(i => (accum(i)._1, ncIR(i + 1))),
+          valueName,
+          seq.indices.map(i => ncIR(i + 1 + accum.length)), ncIR.last)
       case ArrayScan(_, _, accumName, valueName, _) =>
         val IndexedSeq(a: IR, zero: IR, body: IR) = newChildren
         ArrayScan(a, zero, accumName, valueName, body)
@@ -214,6 +221,7 @@ object Copy {
       case x@ApplyIR(fn, args) =>
         val r = ApplyIR(fn, newChildren.map(_.asInstanceOf[IR]))
         r.conversion = x.conversion
+        r.inline = x.inline
         r
       case Apply(fn, args, t) =>
         Apply(fn, newChildren.map(_.asInstanceOf[IR]), t)