--- a/hail/src/main/scala/is/hail/expr/ir/FoldConstants.scala
+++ b/hail/src/main/scala/is/hail/expr/ir/FoldConstants.scala
@@ -28,6 +28,7 @@ object FoldConstants {
              _: NDArrayAgg |
              _: NDArrayWrite |
              _: NDArrayMatMul |
+             _: ArrayFold2 |
              _: Die => None
         case ir: IR if !IsConstant(ir) &&
           Interpretable(ir) &&