… hail/python/hail/ir/register_functions.py
… hail/python/hailtop/hailctl/dev/benchmark/run/table_benchmarks.py
… hail/src/main/scala/is/hail/expr/ir/package.scala
… hail/src/test/scala/is/hail/expr/ir/IRSuite.scala
… hail/src/test/scala/is/hail/expr/ir/SetFunctionsSuite.scala