@tpoterba 2019-09-03
Ported `ArrayFunctions.mean` to ArrayFold2.

Benchmark:
```python
@benchmark
def table_range_means():
    ht = hl.utils.range_table(10_000_000, 16)
    ht = ht.annotate(m = hl.mean(hl.range(0, ht.idx % 1111)))
    ht._force_count()
```

Master:
```
2019-09-03 09:39:05,777: INFO: [1/1] Running table_range_means...
2019-09-03 09:40:52,557: INFO:     burn in: 106.78s
2019-09-03 09:42:34,333: INFO:     run 1: 101.78s
2019-09-03 09:44:14,982: INFO:     run 2: 100.65s
2019-09-03 09:45:53,590: INFO:     run 3: 98.61s
```

PR:
```
2019-09-03 09:47:26,110: INFO: [1/1] Running table_range_means...
2019-09-03 09:47:29,465: INFO:     burn in: 3.35s
2019-09-03 09:47:32,615: INFO:     run 1: 3.15s
2019-09-03 09:47:35,703: INFO:     run 2: 3.09s
2019-09-03 09:47:38,840: INFO:     run 3: 3.14s
```