@tpoterba 2019-09-03
parent PR is in, ready for review.

Note that the next step is replacing ArrayFold with ArrayFold2.